Move renderMorphologyMeasurement to entity renderer module
