+++ src/util/__tests__/listing-selectors.test.ts
@@ -10,7 +10,6 @@ import {
   selectorFnSpecies,
   selectorFnStatistic,
   selectorFnSynaptic,
-  renderMorphologyMeasurement,
 } from '@/util/explore-section/listing-selectors';
 import { SynapticPosition, SynapticType } from '@/types/explore-section/misc';
 import { formatNumber } from '@/util/common';
@@ -550,149 +549,6 @@ describe('Selectors', () => {
         },
         SynapticPosition.Post,
         SynapticType.CellType
-      );
-      expect(result).toBe(DisplayMessages.NO_DATA_STRING);
-    });
-  });
-
-  describe('selectorFnMorphologyFeature', () => {
-    const mockSource = {
-      featureSeries: [
-        {
-          compartment: 'NeuronMorphology',
-          label: 'Total Depth',
-          statistic: 'mean',
-          unit: 'dimensionless',
-          value: 853.0,
-        },
-        {
-          compartment: 'Axon',
-          label: 'Partition Asymmetry',
-          statistic: 'standard deviation',
-          unit: 'dimensionless',
-          value: 0.34878757070556377,
-        },
-        {
-          compartment: 'NeuronMorphology',
-          label: 'Total Length',
-          statistic: 'mean',
-          unit: 'μm',
-          value: 13944.989401578903,
-        },
-        {
-          compartment: 'Axon',
-          label: 'Section Strahler Orders',
-          statistic: 'maximum',
-          unit: 'dimensionless',
-          value: 5.0,
-        },
-        {
-          compartment: 'Axon',
-          label: 'Total Length',
-          statistic: 'maximum',
-          unit: 'μm',
-          value: 9600.45951640606,
-        },
-        {
-          compartment: 'Soma',
-          label: 'Soma Radius',
-          statistic: 'minimum',
-          unit: 'μm',
-          value: 5.46871709813711,
-        },
-        {
-          compartment: 'BasalDendrite',
-          label: 'Total Length',
-          statistic: 'mean',
-          unit: 'μm',
-          value: 4344.529885172844,
-        },
-        {
-          compartment: 'BasalDendrite',
-          label: 'Section Strahler Orders',
-          statistic: 'standard deviation',
-          unit: 'dimensionless',
-          value: 0.8123269504088237,
-        },
-        {
-          compartment: 'ApicalDendrite',
-          label: 'Total Length',
-          statistic: 'minimum',
-          unit: 'μm',
-          value: 0.0,
-        },
-        {
-          compartment: 'ApicalDendrite',
-          label: 'Total Length',
-          statistic: 'maximum',
-          unit: 'μm',
-          value: '0.0',
-        },
-      ],
-    };
-
-    it('returns formatted statistic value if present', () => {
-      const result = renderMorphologyMeasurement(
-        // @ts-ignore
-        mockSource,
-        'NeuronMorphology',
-        'Total Length',
-        'mean',
-        true
-      );
-      expect(result).toBe(`${formatNumber(13944.989401578903)} μm`);
-    });
-
-    it('returns formatted statistic value without unit if showUnits is false', () => {
-      const result = renderMorphologyMeasurement(
-        // @ts-ignore
-        mockSource,
-        'NeuronMorphology',
-        'Total Length',
-        'mean',
-        false
-      );
-      expect(result).toBe(formatNumber(13944.989401578903));
-    });
-
-    it('returns double the value for Soma Radius', () => {
-      const result = renderMorphologyMeasurement(
-        // @ts-ignore
-        mockSource,
-        'Soma',
-        'Soma Radius',
-        'minimum',
-        true
-      );
-      expect(result).toBe(`${formatNumber(10.94)} μm`);
-    });
-
-    it('returns NO_DATA_STRING if statistic is not found', () => {
-      // @ts-ignore
-      const result = renderMorphologyMeasurement(mockSource, 'compartment', 'nonexistent', 'mean');
-      expect(result).toBe(DisplayMessages.NO_DATA_STRING);
-    });
-
-    it('returns NO_DATA_STRING for 0 value', () => {
-      const result = renderMorphologyMeasurement(
-        // @ts-ignore
-        mockSource,
-        'ApicalDendrite',
-        'Total Length',
-        'minimum',
-        true
-      );
-      expect(result).toBe(DisplayMessages.NO_DATA_STRING);
-    });
-
-    it('returns NO DATA STRING when value is not a number', () => {
-      const result = renderMorphologyMeasurement(
-        // @ts-ignore
-        mockSource,
-        'ApicalDendrite',
-        'Total Length',
-        'maximum',
-        true
       );
       expect(result).toBe(DisplayMessages.NO_DATA_STRING);
     });
