+++ src/util/explore-section/listing-selectors.ts
@@ -171,43 +171,3 @@ export const selectorFnSynaptic = (
   const preSynaptic = synapticList.find((synaptic) => synaptic.about === type);
   return preSynaptic?.label ?? DisplayMessages.NO_DATA_STRING;
 };
-
-/**
- * Renders a specific morphology measurement
- *
- * @param {IReconstructionMorphologyExpanded} morphology
- * @param {string} structuralDomain - The compartment to serialize.
- * @param {string} label - The label to serialize.
- * @param {string} measurementType - The statistic to serialize.
- * @param {boolean} showUnits - Whether to show the units.
- *
- * @returns {string} - The rendered text value.
- */
-export const renderMorphologyMeasurement = (
-  morphology: IReconstructionMorphologyExpanded | IReconstructionMorphology,
-  structuralDomain: string,
-  label: string,
-  measurementType: string,
-  showUnits?: boolean
-): ReactNode => {
-  if (!morphology || !('measurement_annotation' in morphology))
-    return DisplayMessages.NO_DATA_STRING;
-
-  const measurementKinds = morphology.measurement_annotation.measurement_kinds;
-
-  const measurementKind = measurementKinds?.find(
-    (mk) => mk.structural_domain === structuralDomain && mk.pref_label === label
-  );
-
-  const measurement = measurementKind?.measurement_items.find((mi) => mi.name === measurementType);
-
-  if (!measurement) return DisplayMessages.NO_DATA_STRING;
-
-  const { unit, value } = measurement;
-  const unitSuffix = showUnits ? `${unit}` : '';
-
-  // TODO: Correct the data if this is still needed.
-  // if (label === 'soma_radius') value = 2 * measurement.value;
-
-  return `${formatNumber(value)}${unitSuffix}`;
-};
